refactor: 공지사항 리스트 네이밍 변경
