+++ src/pages/BoardPage/Notice/Notice.module.scss
@@ -44,12 +44,3 @@
     margin: 0;
   }
 }
-
-.post-loading {
-  width: 835px;
-  height: 690px;
-  background-color: white;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
