+++ src/pages/BoardPage/Notice/hooks/usePageParams.ts
@@ -1,11 +1,7 @@
 import useParamsHandler from 'utils/hooks/useParamsHandler';
 
 const usePageParams = () => {
-  const { params, setParams } = useParamsHandler();
-
-  if (params.page === undefined) {
-    setParams('page', '1', { deleteBeforeParam: false, replacePage: true });
-  }
+  const { params } = useParamsHandler();
 
   return params.page ?? '1';
 };
