+++ src/pages/BoardPage/Notice/index.tsx
@@ -1,4 +1,6 @@
-import useHotArticleList from 'pages/BoardPage/Notice/hooks/useHotPost';
+import { Suspense } from 'react';
+import useHotArticleList from 'pages/BoardPage/Notice/hooks/useHotArticle';
+import { ReactComponent as LoadingSpinner } from 'assets/svg/loading-spinner.svg';
 import { Outlet } from 'react-router-dom';
 import styles from './Notice.module.scss';
 
@@ -10,7 +12,9 @@ function Notice() {
         <div className={styles.header}>
           <h1 className={styles.header__title}>공지사항</h1>
         </div>
-        <Outlet />
+        <Suspense fallback={<LoadingSpinner className={styles['content__loading-spinner']} />}>
+          <Outlet />
+        </Suspense>
       </div>
       { hotArticleList }
     </div>
